gif_bot/requirements.txt
aiogram==3.7.0
aiofiles==23.2.1
python-dotenv==1.0.0
ffmpeg-python==0.2.0
